+++ example/lib/main.dart
@@ -1,7 +1,7 @@
-import 'package:custom_sliding_segmented_control/custom_sliding_segmented_control.dart';
 import 'package:example/example_1.dart';
 import 'package:example/example_10.dart';
 import 'package:example/example_11.dart';
+import 'package:example/example_12.dart';
 import 'package:example/example_2.dart';
 import 'package:example/example_3.dart';
 import 'package:example/example_4.dart';
@@ -48,6 +48,8 @@ class _MyHomePageState extends State<MyHomePage> {
           mainAxisAlignment: MainAxisAlignment.center,
           crossAxisAlignment: CrossAxisAlignment.center,
           children: [
+            const Example12(),
+            const SizedBox(height: 16),
             const Example3(),
             const SizedBox(height: 16),
             const Example1(),