+++ lib/custom_sliding_segmented_control.dart
@@ -1,5 +1,6 @@
 library custom_sliding_segmented_control;
 
 export 'src/custom_segmented_controller.dart';
+export 'src/divider_settings.dart';
 export 'src/segment_settings.dart';
 export 'src/widget.dart';